:bug: fixes promise resolution in transform_entries
This bug was able to slip by since the test was assuming a mutative strategy for the transform function. A nasty bug arises when you chose to NOT mutate the data and instead make copies because the original data is returned, meaning the `map` is not behaving as a map but instead a simple `forEach` loop. This resolves this. I can create a failing test case if that would be helpful.